vsprintf: kptr_restrict is okay in IRQ when 2
The kptr_restrict flag, when set to 1, only prints the kernel address when
the user has CAP_SYSLOG.  When it is set to 2, the kernel address is
always printed as zero.  When set to 1, this needs to check whether or not
we're in IRQ.  However, when set to 2, this check is unneccessary, and
produces confusing results in dmesg.  Thus, only make sure we're not in
IRQ when mode 1 is used, but not mode 2.

[akpm@linux-foundation.org: coding-style fixes]
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Cc: Rasmus Villemoes <linux@rasmusvillemoes.dk>
Cc: Kees Cook <keescook@chromium.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>