mm: thp: fix SMP race condition between THP page fault and MADV_DONTNEED
pmd_trans_unstable()/pmd_none_or_trans_huge_or_clear_bad() were introduced
to locklessy (but atomically) detect when a pmd is a regular (stable) pmd
or when the pmd is unstable and can infinitely transition from pmd_none()
and pmd_trans_huge() from under us, while only holding the mmap_sem for
reading (for writing not).

While holding the mmap_sem only for reading, MADV_DONTNEED can run from
under us and so before we can assume the pmd to be a regular stable pmd we
need to compare it against pmd_none() and pmd_trans_huge() in an atomic
way, with pmd_trans_unstable().  The old pmd_trans_huge() left a tiny
window for a race.

Useful applications are unlikely to notice the difference as doing
MADV_DONTNEED concurrently with a page fault would lead to undefined
behavior.

[akpm@linux-foundation.org: tidy up comment grammar/layout]
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Reported-by: Kirill A. Shutemov <kirill.shutemov@linux.intel.com>
Cc: <stable@vger.kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>