proc: /proc/<pid>/timerslack_ns permissions fixes
This patch adjusts the timerslack_ns file permissions to be 0666 but
requires PTRACE_MODE_ATTACH_FSCREDS to read or write the value.

This allows tasks with sufficient privledges (CAP_SYS_PTRACE) to be able
to modify a the timerslack for proccesses owned by a different user.

This patch also fixes a return value from EINVAL to EPERM, and does task
locking consistently, given we're handling u64s on 32bit systems.  It also
makes use of kstrtoull_from_user which simplifies some code.

I've also wired up the Android userspace side to use this interface, and
tested it there, and things seem to be working properly ( - with some
selinux noise, I still need to figure out the selinux policy changes, but
its working with permissive mode).

Signed-off-by: John Stultz <john.stultz@linaro.org>
Cc: Arjan van de Ven <arjan@linux.intel.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Oren Laadan <orenl@cellrox.com>
Cc: Ruchi Kandoi <kandoiruchi@google.com>
Cc: Rom Lemarchand <romlem@android.com>
Cc: Kees Cook <keescook@chromium.org>
Cc: Android Kernel Team <kernel-team@android.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>