mm: thp: Set THP defrag by default to madvise and add a stall-free defrag option -fix
The following is a fix to the patch
mm-thp-set-thp-defrag-by-default-to-madvise-and-add-a-stall-free-defrag-option.patch
based on feedback from Vlastimil Babka.  It removes an unnecessary
VM_BUG_ON for tidyness, clarifies documentation and adds a check
forbidding someone writing "defer" to the enable knob for transparent huge
pages.  The ack from Vlastimil applies for the original patch plus this
fix combined.

Signed-off-by: Mel Gorman <mgorman@techsingularity.net>
Acked-by: Rik van Riel <riel@redhat.com>
Acked-by: Johannes Weiner <hannes@cmpxchg.org>
Acked-by: Vlastimil Babka <vbabka@suse.cz>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>